Time: 15 ms (10.47%), Space: 8.6 MB (13.23%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -3,7 +3,7 @@ public:
     int mySqrt(int x) {
         int result=0;
         for(int i=1;i<=x; i++){
-            if((long long)i*i<=x) {
+            if((double)i*i<=x) {
                 result= i;
             }
             else{
